Create mock local storage object
